Rename PiGui -> PiGui::Instance
Make PiGui a namespace for easier method scoping, allows us to be more flexible

PiGui::Instance only has the bare minimum of methods needed to run ImGui

Move PiGui RegisterClass call to pigui/PiGuiLua - we'll further unify LuaPiGui
and PiGuiLua later.
