+++ src/lua/Lua.cpp
@@ -107,8 +107,6 @@ namespace Lua {
 		GameUI::Lua::Init();
 		SceneGraph::Lua::Init();
 
-		PiGUI::Lua::Init();
-
 		// XXX load everything. for now, just modules
 		lua_State *l = Lua::manager->GetLuaState();
 		pi_lua_dofile(l, "libs/autoload.lua");
