+++ src/pigui/PiGui.cpp
@@ -5,6 +5,8 @@
 #include "Input.h"
 #include "Pi.h"
 
+#include "graphics/Texture.h"
+#include "graphics/opengl/RendererGL.h"
 #include "graphics/opengl/TextureGL.h" // nasty, usage of GL is implementation specific
 #include "imgui/imgui.h"
 
@@ -12,9 +14,6 @@
 #define IMGUI_IMPL_OPENGL_LOADER_GLEW 1
 #include "imgui/examples/imgui_impl_opengl3.h"
 #include "imgui/examples/imgui_impl_sdl.h"
-// to get ImVec2 + ImVec2
-#define IMGUI_DEFINE_MATH_OPERATORS true
-#include "imgui/imgui_internal.h"
 
 #include <float.h>
 #include <stdio.h>
@@ -24,7 +23,37 @@
 #define NANOSVGRAST_IMPLEMENTATION
 #include "nanosvg/nanosvgrast.h"
 
-std::vector<Graphics::Texture *> PiGui::m_svg_textures;
+using namespace PiGui;
+
+std::vector<Graphics::Texture *> m_svg_textures;
+
+std::vector<Graphics::Texture *> &PiGui::GetSVGTextures()
+{
+	return m_svg_textures;
+}
+
+static void *makeTexture(unsigned char *pixels, int width, int height)
+{
+	PROFILE_SCOPED()
+	// this is not very pretty code and uses the Graphics::TextureGL class directly
+	// Texture descriptor defines the size, type.
+	// Gone for LINEAR_CLAMP here and RGBA like the original code
+	const vector2f texSize(1.0f, 1.0f);
+	const vector3f dataSize(width, height, 0.0f);
+	const Graphics::TextureDescriptor texDesc(Graphics::TEXTURE_RGBA_8888,
+		dataSize, texSize, Graphics::LINEAR_CLAMP,
+		false, false, false, 0, Graphics::TEXTURE_2D);
+	// Create the texture, calling it via renderer directly avoids the caching call of TextureBuilder
+	// However interestingly this gets called twice which would have been a WIN for the TextureBuilder :/
+	Graphics::Texture *pTex = Pi::renderer->CreateTexture(texDesc);
+	// Update it with the actual pixels, this is a two step process due to legacy code
+	pTex->Update(pixels, dataSize, Graphics::TEXTURE_RGBA_8888);
+	// nasty bit as I invoke the TextureGL
+	Graphics::OGL::TextureGL *pGLTex = reinterpret_cast<Graphics::OGL::TextureGL *>(pTex);
+	Uint32 result = pGLTex->GetTextureID();
+	PiGui::GetSVGTextures().push_back(pTex); // store for cleanup later
+	return reinterpret_cast<void *>(result);
+}
 
 ImTextureID PiGui::RenderSVG(std::string svgFilename, int width, int height)
 {
@@ -85,7 +114,41 @@ ImTextureID PiGui::RenderSVG(std::string svgFilename, int width, int height)
 	return makeTexture(img, W, H);
 }
 
-ImFont *PiGui::GetFont(const std::string &name, int size)
+//
+//	PiGui::Instance
+//
+
+Instance::Instance() :
+	m_should_bake_fonts(true)
+{
+	// TODO: clang-format doesn't like list initializers inside function calls
+	// clang-format off
+	PiFont uiheading("orbiteer", {
+		PiFace("DejaVuSans.ttf", /*18.0/20.0*/ 1.2),
+		PiFace("wqy-microhei.ttc", 1.0),
+		PiFace("Orbiteer-Bold.ttf", 1.0) // imgui only supports 0xffff, not 0x10ffff
+	});
+	AddFontDefinition(uiheading);
+
+	PiFont guifont("pionillium", {
+		PiFace("DejaVuSans.ttf", 13.0 / 14.0),
+		PiFace("wqy-microhei.ttc", 1.0),
+		PiFace("PionilliumText22L-Medium.ttf", 1.0)
+	});
+	AddFontDefinition(guifont);
+	// clang-format on
+
+	// Output("Fonts:\n");
+	for (auto entry : m_font_definitions) {
+		//		Output("  entry %s:\n", entry.first.c_str());
+		entry.second.describe();
+	}
+
+	// ensure the tooltip font exists
+	GetFont("pionillium", 14);
+};
+
+ImFont *Instance::GetFont(const std::string &name, int size)
 {
 	PROFILE_SCOPED()
 	auto iter = m_fonts.find(std::make_pair(name, size));
@@ -97,7 +160,7 @@ ImFont *PiGui::GetFont(const std::string &name, int size)
 	return font;
 }
 
-void PiGui::AddGlyph(ImFont *font, unsigned short glyph)
+void Instance::AddGlyph(ImFont *font, unsigned short glyph)
 {
 	PROFILE_SCOPED()
 	// range glyph..glyph
@@ -122,7 +185,7 @@ void PiGui::AddGlyph(ImFont *font, unsigned short glyph)
 	Error("No face in font %s handles glyph %i\n", pifont.name().c_str(), glyph);
 }
 
-ImFont *PiGui::AddFont(const std::string &name, int size)
+ImFont *Instance::AddFont(const std::string &name, int size)
 {
 	PROFILE_SCOPED()
 	auto iter = m_font_definitions.find(name);
@@ -146,7 +209,7 @@ ImFont *PiGui::AddFont(const std::string &name, int size)
 	return m_fonts[std::make_pair(name, size)];
 }
 
-void PiGui::RefreshFontsTexture()
+void Instance::RefreshFontsTexture()
 {
 	PROFILE_SCOPED()
 	// TODO: fix this, do the right thing, don't just re-create *everything* :)
@@ -160,7 +223,7 @@ void PiDefaultStyle(ImGuiStyle &style)
 	style.WindowBorderSize = 0.0f; // Thickness of border around windows. Generally set to 0.0f or 1.0f. Other values not well tested.
 }
 
-void PiGui::Init(SDL_Window *window)
+void Instance::Init(SDL_Window *window)
 {
 	PROFILE_SCOPED()
 
@@ -196,146 +259,14 @@ void PiGui::Init(SDL_Window *window)
 	io.IniFilename = ioIniFilename;
 }
 
-int PiGui::RadialPopupSelectMenu(const ImVec2 &center, std::string popup_id, int mouse_button, std::vector<ImTextureID> tex_ids, std::vector<std::pair<ImVec2, ImVec2>> uvs, unsigned int size, std::vector<std::string> tooltips)
-{
-	PROFILE_SCOPED()
-	// return:
-	// 0 - n for item selected
-	// -1 for nothing chosen, but menu open
-	// -2 for menu closed without an icon chosen
-	// -3 for menu not open
-	int ret = -3;
-
-	// FIXME: Missing a call to query if Popup is open so we can move the PushStyleColor inside the BeginPopupBlock (e.g. IsPopupOpen() in imgui.cpp)
-	// FIXME: Our PathFill function only handle convex polygons, so we can't have items spanning an arc too large else inner concave edge artifact is too visible, hence the ImMax(7,items_count)
-	ImGui::PushStyleColor(ImGuiCol_WindowBg, ImVec4(0, 0, 0, 0));
-	ImGui::PushStyleColor(ImGuiCol_PopupBg, ImVec4(0, 0, 0, 0));
-	ImGui::PushStyleColor(ImGuiCol_Border, ImVec4(0, 0, 0, 0));
-	if (ImGui::BeginPopup(popup_id.c_str())) {
-		ret = -1;
-		const ImVec2 drag_delta = ImVec2(ImGui::GetIO().MousePos.x - center.x, ImGui::GetIO().MousePos.y - center.y);
-		const float drag_dist2 = drag_delta.x * drag_delta.x + drag_delta.y * drag_delta.y;
-
-		const ImGuiStyle &style = ImGui::GetStyle();
-		const float RADIUS_MIN = 20.0f;
-		const float RADIUS_MAX = 90.0f;
-		const float RADIUS_INTERACT_MIN = 20.0f;
-		const int ITEMS_MIN = 4;
-		const float border_inout = 12.0f;
-		const float border_thickness = 4.0f;
-		ImDrawList *draw_list = ImGui::GetWindowDrawList();
-		draw_list->PushClipRectFullScreen();
-		draw_list->PathArcTo(center, (RADIUS_MIN + RADIUS_MAX) * 0.5f, 0.0f, IM_PI * 2.0f * 0.99f, 64); // FIXME: 0.99f look like full arc with closed thick stroke has a bug now
-		draw_list->PathStroke(ImColor(18, 44, 67, 210), true, RADIUS_MAX - RADIUS_MIN);
-
-		const float item_arc_span = 2 * IM_PI / ImMax<int>(ITEMS_MIN, tex_ids.size());
-		float drag_angle = atan2f(drag_delta.y, drag_delta.x);
-		if (drag_angle < -0.5f * item_arc_span)
-			drag_angle += 2.0f * IM_PI;
-
-		int item_hovered = -1;
-		int item_n = 0;
-		for (ImTextureID tex_id : tex_ids) {
-			const char *tooltip = tooltips.at(item_n).c_str();
-			const float inner_spacing = style.ItemInnerSpacing.x / RADIUS_MIN / 2;
-			const float item_inner_ang_min = item_arc_span * (item_n - 0.5f + inner_spacing);
-			const float item_inner_ang_max = item_arc_span * (item_n + 0.5f - inner_spacing);
-			const float item_outer_ang_min = item_arc_span * (item_n - 0.5f + inner_spacing * (RADIUS_MIN / RADIUS_MAX));
-			const float item_outer_ang_max = item_arc_span * (item_n + 0.5f - inner_spacing * (RADIUS_MIN / RADIUS_MAX));
-
-			bool hovered = false;
-			if (drag_dist2 >= RADIUS_INTERACT_MIN * RADIUS_INTERACT_MIN) {
-				if (drag_angle >= item_inner_ang_min && drag_angle < item_inner_ang_max)
-					hovered = true;
-			}
-			bool selected = false;
-
-			int arc_segments = static_cast<int>((64 * item_arc_span / (2 * IM_PI))) + 1;
-			draw_list->PathArcTo(center, RADIUS_MAX - border_inout, item_outer_ang_min, item_outer_ang_max, arc_segments);
-			draw_list->PathArcTo(center, RADIUS_MIN + border_inout, item_inner_ang_max, item_inner_ang_min, arc_segments);
-
-			draw_list->PathFillConvex(hovered ? ImColor(102, 147, 189) : selected ? ImColor(48, 81, 111) : ImColor(48, 81, 111));
-			if (hovered) {
-				// draw outer / inner extra segments
-				draw_list->PathArcTo(center, RADIUS_MAX - border_thickness, item_outer_ang_min, item_outer_ang_max, arc_segments);
-				draw_list->PathStroke(ImColor(102, 147, 189), false, border_thickness);
-				draw_list->PathArcTo(center, RADIUS_MIN + border_thickness, item_outer_ang_min, item_outer_ang_max, arc_segments);
-				draw_list->PathStroke(ImColor(102, 147, 189), false, border_thickness);
-			}
-			ImVec2 text_size = ImVec2(size, size);
-			ImVec2 text_pos = ImVec2(
-				center.x + cosf((item_inner_ang_min + item_inner_ang_max) * 0.5f) * (RADIUS_MIN + RADIUS_MAX) * 0.5f - text_size.x * 0.5f,
-				center.y + sinf((item_inner_ang_min + item_inner_ang_max) * 0.5f) * (RADIUS_MIN + RADIUS_MAX) * 0.5f - text_size.y * 0.5f);
-			draw_list->AddImage(tex_id, text_pos, ImVec2(text_pos.x + size, text_pos.y + size), uvs[item_n].first, uvs[item_n].second);
-			ImGui::SameLine();
-			if (hovered) {
-				item_hovered = item_n;
-				ImGui::SetTooltip("%s", tooltip);
-			}
-			item_n++;
-		}
-		draw_list->PopClipRect();
-
-		if (ImGui::IsMouseReleased(mouse_button)) {
-			ImGui::CloseCurrentPopup();
-			if (item_hovered == -1)
-				ret = -2;
-			else
-				ret = item_hovered;
-		}
-		ImGui::EndPopup();
-	} else {
-		// Output("WARNING: RadialPopupSelectMenu BeginPopup failed: %s\n", popup_id.c_str());
-	}
-	ImGui::PopStyleColor(3);
-	return ret;
-}
-
-bool PiGui::CircularSlider(const ImVec2 &center, float *v, float v_min, float v_max)
-{
-	PROFILE_SCOPED()
-	ImDrawList *draw_list = ImGui::GetWindowDrawList();
-	ImGuiWindow *window = ImGui::GetCurrentWindow();
-	const ImGuiID id = window->GetID("circularslider");
-	draw_list->AddCircle(center, 17, ImColor(100, 100, 100), 128, 12.0);
-	draw_list->PathArcTo(center, 17, 0, M_PI * 2.0 * (*v - v_min) / (v_max - v_min), 64);
-	draw_list->PathStroke(ImColor(200, 200, 200), false, 12.0);
-	ImRect grab_bb;
-	return ImGui::SliderBehavior(ImRect(center.x - 17, center.y - 17, center.x + 17, center.y + 17),
-		id, ImGuiDataType_Float, v, &v_min, &v_max, "%.4f", 1.0, ImGuiSliderFlags_None, &grab_bb);
-}
-
-bool PiGui::ProcessEvent(SDL_Event *event)
+bool Instance::ProcessEvent(SDL_Event *event)
 {
 	PROFILE_SCOPED()
 	ImGui_ImplSDL2_ProcessEvent(event);
 	return false;
 }
 
-void *PiGui::makeTexture(unsigned char *pixels, int width, int height)
-{
-	PROFILE_SCOPED()
-	// this is not very pretty code and uses the Graphics::TextureGL class directly
-	// Texture descriptor defines the size, type.
-	// Gone for LINEAR_CLAMP here and RGBA like the original code
-	const vector2f texSize(1.0f, 1.0f);
-	const vector3f dataSize(width, height, 0.0f);
-	const Graphics::TextureDescriptor texDesc(Graphics::TEXTURE_RGBA_8888,
-		dataSize, texSize, Graphics::LINEAR_CLAMP,
-		false, false, false, 0, Graphics::TEXTURE_2D);
-	// Create the texture, calling it via renderer directly avoids the caching call of TextureBuilder
-	// However interestingly this gets called twice which would have been a WIN for the TextureBuilder :/
-	Graphics::Texture *pTex = Pi::renderer->CreateTexture(texDesc);
-	// Update it with the actual pixels, this is a two step process due to legacy code
-	pTex->Update(pixels, dataSize, Graphics::TEXTURE_RGBA_8888);
-	// nasty bit as I invoke the TextureGL
-	Graphics::OGL::TextureGL *pGLTex = reinterpret_cast<Graphics::OGL::TextureGL *>(pTex);
-	Uint32 result = pGLTex->GetTextureID();
-	m_svg_textures.push_back(pTex); // store for cleanup later
-	return reinterpret_cast<void *>(result);
-}
-
-void PiGui::NewFrame(SDL_Window *window)
+void Instance::NewFrame(SDL_Window *window)
 {
 	PROFILE_SCOPED()
 
@@ -361,7 +292,7 @@ void PiGui::NewFrame(SDL_Window *window)
 	ImGui::SetMouseCursor(ImGuiMouseCursor_Arrow);
 }
 
-void PiGui::EndFrame()
+void Instance::EndFrame()
 {
 	PROFILE_SCOPED()
 
@@ -388,7 +319,7 @@ void PiGui::EndFrame()
 	}
 }
 
-void PiGui::Render()
+void Instance::Render()
 {
 	PROFILE_SCOPED()
 	EndFrame();
@@ -405,7 +336,7 @@ void PiGui::Render()
 	}
 }
 
-void PiGui::ClearFonts()
+void Instance::ClearFonts()
 {
 	PROFILE_SCOPED()
 	ImGuiIO &io = ImGui::GetIO();
@@ -415,7 +346,7 @@ void PiGui::ClearFonts()
 	io.Fonts->Clear();
 }
 
-void PiGui::BakeFont(PiFont &font)
+void Instance::BakeFont(PiFont &font)
 {
 	PROFILE_SCOPED()
 	ImGuiIO &io = ImGui::GetIO();
@@ -456,7 +387,7 @@ void PiGui::BakeFont(PiFont &font)
 	imfont->MissingGlyphs.clear();
 }
 
-void PiGui::BakeFonts()
+void Instance::BakeFonts()
 {
 	PROFILE_SCOPED()
 	//	Output("Baking fonts\n");
@@ -483,192 +414,7 @@ void PiGui::BakeFonts()
 	RefreshFontsTexture();
 }
 
-static void drawThrust(ImDrawList *draw_list, const ImVec2 &center, const ImVec2 &up, float value, const ImColor &fg, const ImColor &bg)
-{
-	PROFILE_SCOPED()
-	float factor = 0.1; // how much to offset from center
-	const ImVec2 step(up.x * 0.5, up.y * 0.5);
-	const ImVec2 left(-step.y * (1.0 - factor), step.x * (1.0 - factor));
-	const ImVec2 u(up.x * (1.0 - factor), up.y * (1.0 - factor));
-	const ImVec2 c(center + ImVec2(u.x * factor, u.y * factor));
-	const ImVec2 right(-left.x, -left.y);
-	const ImVec2 leftmiddle = c + step + left;
-	const ImVec2 rightmiddle = c + step + right;
-	const ImVec2 bb_lowerright = c + right;
-	const ImVec2 bb_upperleft = c + left + ImVec2(u.x * value, u.y * value);
-	const ImVec2 lefttop = c + u + left;
-	const ImVec2 righttop = c + u + right;
-	const ImVec2 minimum(fmin(bb_upperleft.x, bb_lowerright.x), fmin(bb_upperleft.y, bb_lowerright.y));
-	const ImVec2 maximum(fmax(bb_upperleft.x, bb_lowerright.x), fmax(bb_upperleft.y, bb_lowerright.y));
-	ImVec2 points[] = { c, leftmiddle, lefttop, righttop, rightmiddle };
-	draw_list->AddConvexPolyFilled(points, 5, bg);
-	draw_list->PushClipRect(minimum - ImVec2(1, 1), maximum + ImVec2(1, 1));
-	draw_list->AddConvexPolyFilled(points, 5, fg);
-	draw_list->PopClipRect();
-}
-
-void PiGui::ThrustIndicator(const std::string &id_string, const ImVec2 &size_arg, const ImVec4 &thrust, const ImVec4 &velocity, const ImVec4 &bg_col, int frame_padding, ImColor vel_fg, ImColor vel_bg, ImColor thrust_fg, ImColor thrust_bg)
-{
-	PROFILE_SCOPED()
-	ImGuiWindow *window = ImGui::GetCurrentWindow();
-	if (window->SkipItems)
-		return;
-
-	ImGuiContext &g = *GImGui;
-	const ImGuiStyle &style = g.Style;
-	const ImGuiID id = window->GetID(id_string.c_str());
-
-	ImVec2 pos = window->DC.CursorPos;
-	// if ((flags & ImGuiButtonFlags_AlignTextBaseLine) && style.FramePadding.y < window->DC.CurrentLineTextBaseOffset) // Try to vertically align buttons that are smaller/have no padding so that text baseline matches (bit hacky, since it shouldn't be a flag)
-	//     pos.y += window->DC.CurrentLineTextBaseOffset - style.FramePadding.y;
-	ImVec2 size = ImGui::CalcItemSize(size_arg, style.FramePadding.x * 2.0f, style.FramePadding.y * 2.0f);
-
-	const ImVec2 padding = (frame_padding >= 0) ? ImVec2(static_cast<float>(frame_padding), static_cast<float>(frame_padding)) : style.FramePadding;
-	const ImRect bb(pos, pos + size + padding * 2);
-	const ImRect inner_bb(pos + padding, pos + padding + size);
-
-	ImGui::ItemSize(bb, style.FramePadding.y);
-	if (!ImGui::ItemAdd(bb, id))
-		return;
-
-	// Render
-	const ImU32 col = ImGui::GetColorU32(static_cast<ImGuiCol>(ImGuiCol_Button));
-	ImGui::RenderFrame(bb.Min, bb.Max, col, true, style.FrameRounding);
-	ImDrawList *draw_list = ImGui::GetWindowDrawList();
-	if (bg_col.w > 0.0f)
-		draw_list->AddRectFilled(inner_bb.Min, inner_bb.Max, ImGui::GetColorU32(bg_col));
-	const ImVec2 leftupper = inner_bb.Min;
-	const ImVec2 rightlower = inner_bb.Max;
-	const ImVec2 rightcenter((rightlower.x - leftupper.x) * 0.8 + leftupper.x, (rightlower.y + leftupper.y) / 2);
-	const ImVec2 leftcenter((rightlower.x - leftupper.x) * 0.35 + leftupper.x, (rightlower.y + leftupper.y) / 2);
-	const ImVec2 up(0, -std::abs(leftupper.y - rightlower.y) * 0.4);
-	const ImVec2 left(-up.y, up.x);
-	float thrust_fwd = fmax(thrust.z, 0);
-	float thrust_bwd = fmax(-thrust.z, 0);
-	float thrust_left = fmax(-thrust.x, 0);
-	float thrust_right = fmax(thrust.x, 0);
-	float thrust_up = fmax(-thrust.y, 0);
-	float thrust_down = fmax(thrust.y, 0);
-	// actual thrust
-	drawThrust(draw_list, rightcenter, up, thrust_fwd, thrust_fg, thrust_bg);
-	drawThrust(draw_list, rightcenter, ImVec2(-up.x, -up.y), thrust_bwd, thrust_fg, thrust_bg);
-	drawThrust(draw_list, leftcenter, up, thrust_up, thrust_fg, thrust_bg);
-	drawThrust(draw_list, leftcenter, ImVec2(-up.x, -up.y), thrust_down, thrust_fg, thrust_bg);
-	drawThrust(draw_list, leftcenter, left, thrust_left, thrust_fg, thrust_bg);
-	drawThrust(draw_list, leftcenter, ImVec2(-left.x, -left.y), thrust_right, thrust_fg, thrust_bg);
-	// forward/back velocity
-	draw_list->AddLine(rightcenter + up, rightcenter - up, vel_bg, 3);
-	draw_list->AddLine(rightcenter, rightcenter - up * velocity.z, vel_fg, 3);
-	// left/right velocity
-	draw_list->AddLine(leftcenter + left, leftcenter - left, vel_bg, 3);
-	draw_list->AddLine(leftcenter, leftcenter + left * velocity.x, vel_fg, 3);
-	// up/down velocity
-	draw_list->AddLine(leftcenter + up, leftcenter - up, vel_bg, 3);
-	draw_list->AddLine(leftcenter, leftcenter + up * velocity.y, vel_fg, 3);
-	// Automatically close popups
-	//if (pressed && !(flags & ImGuiButtonFlags_DontClosePopups) && (window->Flags & ImGuiWindowFlags_Popup))
-	//    CloseCurrentPopup();
-}
-
-bool PiGui::LowThrustButton(const char *id_string, const ImVec2 &size_arg, int thrust_level, const ImVec4 &bg_col, int frame_padding, ImColor gauge_fg, ImColor gauge_bg)
-{
-	PROFILE_SCOPED()
-	std::string label = std::to_string(thrust_level);
-	ImGuiWindow *window = ImGui::GetCurrentWindow();
-	if (window->SkipItems)
-		return false;
-
-	ImGuiContext &g = *GImGui;
-	const ImGuiStyle &style = g.Style;
-	const ImGuiID id = window->GetID(id_string);
-	const ImVec2 label_size = ImGui::CalcTextSize(label.c_str(), NULL, true);
-
-	ImVec2 pos = window->DC.CursorPos;
-	// if ((flags & ImGuiButtonFlags_AlignTextBaseLine) && style.FramePadding.y < window->DC.CurrentLineTextBaseOffset) // Try to vertically align buttons that are smaller/have no padding so that text baseline matches (bit hacky, since it shouldn't be a flag)
-	//     pos.y += window->DC.CurrentLineTextBaseOffset - style.FramePadding.y;
-	ImVec2 size = ImGui::CalcItemSize(size_arg, label_size.x + style.FramePadding.x * 2.0f, label_size.y + style.FramePadding.y * 2.0f);
-
-	const ImVec2 padding = (frame_padding >= 0) ? ImVec2(static_cast<float>(frame_padding), static_cast<float>(frame_padding)) : style.FramePadding;
-	const ImRect bb(pos, pos + size + padding * 2);
-	const ImRect inner_bb(pos + padding, pos + padding + size);
-
-	ImGui::ItemSize(bb, style.FramePadding.y);
-	if (!ImGui::ItemAdd(bb, id))
-		return false;
-
-	// if (window->DC.ButtonRepeat) flags |= ImGuiButtonFlags_Repeat;
-	bool hovered, held;
-	bool pressed = ImGui::ButtonBehavior(bb, id, &hovered, &held, 0); // flags
-
-	// Render
-	const ImU32 col = ImGui::GetColorU32(static_cast<ImGuiCol>((hovered && held) ? ImGuiCol_ButtonActive : (hovered ? ImGuiCol_ButtonHovered : ImGuiCol_Button)));
-	ImGui::RenderFrame(bb.Min, bb.Max, col, true, style.FrameRounding);
-	const ImVec2 center = (inner_bb.Min + inner_bb.Max) / 2;
-	float radius = (inner_bb.Max.x - inner_bb.Min.x) * 0.4;
-	float thickness = 4;
-	ImDrawList *draw_list = ImGui::GetWindowDrawList();
-	if (bg_col.w > 0.0f)
-		draw_list->AddRectFilled(inner_bb.Min, inner_bb.Max, ImGui::GetColorU32(bg_col));
-
-	draw_list->PathArcTo(center, radius, 0, IM_PI * 2, 16);
-	draw_list->PathStroke(gauge_bg, false, thickness);
-
-	draw_list->PathArcTo(center, radius, IM_PI, IM_PI + IM_PI * 2 * (thrust_level / 100.0), 16);
-	draw_list->PathStroke(gauge_fg, false, thickness);
-	ImGui::RenderTextClipped(bb.Min + style.FramePadding, bb.Max - style.FramePadding, label.c_str(), NULL, &label_size, style.ButtonTextAlign, &bb);
-
-	// Automatically close popups
-	//if (pressed && !(flags & ImGuiButtonFlags_DontClosePopups) && (window->Flags & ImGuiWindowFlags_Popup))
-	//    CloseCurrentPopup();
-
-	return pressed;
-}
-
-// frame_padding < 0: uses FramePadding from style (default)
-// frame_padding = 0: no framing
-// frame_padding > 0: set framing size
-// The color used are the button colors.
-bool PiGui::ButtonImageSized(ImTextureID user_texture_id, const ImVec2 &size, const ImVec2 &imgSize, const ImVec2 &uv0, const ImVec2 &uv1, int frame_padding, const ImVec4 &bg_col, const ImVec4 &tint_col)
-{
-	ImGuiWindow *window = ImGui::GetCurrentWindow();
-	if (window->SkipItems)
-		return false;
-
-	ImGuiContext &g = *GImGui;
-	const ImGuiStyle &style = g.Style;
-
-	// Default to using texture ID as ID. User can still push string/integer prefixes.
-	// We could hash the size/uv to create a unique ID but that would prevent the user from animating UV.
-	ImGui::PushID((void *)user_texture_id);
-	const ImGuiID id = window->GetID("#image");
-	ImGui::PopID();
-
-	ImVec2 imgPadding = (size - imgSize) / 2;
-	imgPadding.x = imgPadding.x < 0 || imgSize.x <= 0 ? 0 : imgPadding.x;
-	imgPadding.y = imgPadding.y < 0 || imgSize.y <= 0 ? 0 : imgPadding.y;
-
-	const ImVec2 padding = (frame_padding >= 0) ? ImVec2((float)frame_padding, (float)frame_padding) : style.FramePadding;
-	const ImRect bb(window->DC.CursorPos, window->DC.CursorPos + size + padding * 2);
-	const ImRect image_bb(window->DC.CursorPos + padding + imgPadding, window->DC.CursorPos + padding + size - imgPadding);
-	ImGui::ItemSize(bb);
-	if (!ImGui::ItemAdd(bb, id))
-		return false;
-
-	bool hovered, held;
-	bool pressed = ImGui::ButtonBehavior(bb, id, &hovered, &held);
-
-	// Render
-	const ImU32 col = ImGui::GetColorU32((held && hovered) ? ImGuiCol_ButtonActive : hovered ? ImGuiCol_ButtonHovered : ImGuiCol_Button);
-	ImGui::RenderNavHighlight(bb, id);
-	ImGui::RenderFrame(bb.Min, bb.Max, col, true, ImClamp((float)ImMin(padding.x, padding.y), 0.0f, style.FrameRounding));
-	if (bg_col.w > 0.0f)
-		window->DrawList->AddRectFilled(image_bb.Min, image_bb.Max, ImGui::GetColorU32(bg_col));
-	window->DrawList->AddImage(user_texture_id, image_bb.Min, image_bb.Max, uv0, uv1, ImGui::GetColorU32(tint_col));
-
-	return pressed;
-}
-
-void PiGui::Uninit()
+void Instance::Uninit()
 {
 	PROFILE_SCOPED()
 	for (auto tex : m_svg_textures) {
@@ -688,25 +434,9 @@ void PiGui::Uninit()
 	ImGui::DestroyContext();
 }
 
-PiGui::PiGui() :
-	m_should_bake_fonts(true)
-{
-	PiFont uiheading("orbiteer", {
-									 PiFace("DejaVuSans.ttf", /*18.0/20.0*/ 1.2), PiFace("wqy-microhei.ttc", 1.0), PiFace("Orbiteer-Bold.ttf", 1.0) // imgui only supports 0xffff, not 0x10ffff
-								 });
-	PiFont guifont("pionillium", { PiFace("DejaVuSans.ttf", 13.0 / 14.0), PiFace("wqy-microhei.ttc", 1.0), PiFace("PionilliumText22L-Medium.ttf", 1.0) });
-	AddFontDefinition(uiheading);
-	AddFontDefinition(guifont);
-
-	// Output("Fonts:\n");
-	for (auto entry : m_font_definitions) {
-		//		Output("  entry %s:\n", entry.first.c_str());
-		entry.second.describe();
-	}
-
-	// ensure the tooltip font exists
-	GetFont("pionillium", 14);
-};
+//
+// PiGui::PiFace
+//
 
 const bool PiFace::isValidGlyph(unsigned short glyph) const
 {
